--- command-usage.txt
@@ -1,16 +1,0 @@
-  0x --cmd <command> [args]
-  0x -c <command> [args]
-
-  Commands
-
-    gen:
-    The gen command will regenerate the flamegraph from
-    a JSON or stacks.out file (autodetected)
-
-    Create a flamegraph from a stacks file (e.g. dtrace
-    and perf script output)
-
-    0x -c gen [flags] <stacks.<pid>.out file>
-
-    Flags include `--tiers` and `--langs`
-
