--- flamegraph.html
@@ -1,1 +1,0 @@
-<meta charset="utf-8"><h1 style="color: rgb(68, 68, 68);">undefined</h1><style>body {padding-left: 2%; background:black}rect:hover {opacity: 0.9}</style><chart></chart><script>undefined</script>
