+++ lib/soft-exit.js
@@ -1,0 +1,1 @@
+process.on('SIGINT', process.exit)
