+++ package.json
@@ -29,9 +29,11 @@
     "jsdom": "^8.0.4",
     "minimist": "^1.2.0",
     "multistream": "^2.1.0",
+    "once": "^1.4.0",
     "open": "0.0.5",
     "perf-sym": "^2.0.0",
     "pump": "^1.0.1",
+    "pumpify": "^1.4.0",
     "single-line-log": "^1.0.1",
     "split2": "^2.0.1",
     "through2": "^2.0.1",
