+++ platform/index.js
@@ -1,0 +1,6 @@
+'use strict'
+const linux = require('./linux')
+const sun = require('./sun')
+const windows = require('./windows')
+
+module.exports = { linux, sun, windows }
