+++ platform/windows.js
@@ -1,0 +1,9 @@
+'use strict'
+
+const { unsupported } = require('../lib/util')
+
+module.exports = windows 
+
+function windows (args, sudo, binary) {
+  unsupported(args)
+}
