--- soft-exit.js
@@ -1,1 +1,0 @@
-process.on('SIGINT', process.exit)
