+++ usage.txt
@@ -1,5 +1,9 @@
 
-  0x [flags] file.js
+  0x [flags] -- [nodeFlags] file.js -- [scriptFlags]  
+
+  --name                  The name of the HTML file, without the .html extension
+                          Can be set to - to write HTML to STDOUT
+                          Default: flamegraph
 
   -o                      Automatically open after finishing
                           Default: false
@@ -7,7 +11,8 @@
   --node=PATH             Path to the node executable to use
                           Default: false
 
-  --node-options=ARGS     List of options to pass to Node
+  --node-options=ARGS     List of flags to pass to Node, can be used
+                          instead of the double dash suffix (--)
 
   --delay | -d            Delay before tracing begins, allows us to ignore
                           initialisation stacks (e.g. module loading).
@@ -19,14 +24,6 @@
 
   --svg                   Generates an `flamegraph.svg` file in the artifact
                           output directory, in addition to the `flamegraph.html` file.
-
-  --preview | -p          Generates an SVG file, prerenders SVG inside HTML
-                          and outputs a PNG to the terminal (if possible)
-                          Depends on imagemagick (brew install imagemagick)
-                          If using iTerm 2.9+ image will be output to terminal
-                          Warning - depending on the amount of stacks this
-                          option can take tens of seconds
-                          Default: false
 
   --langs | -l            Coarser form of tiers, colour code by JS or C
                           Default: false
@@ -53,13 +50,20 @@
   --trace-info            Show output from dtrace or perf tools
                           Default: false
 
-  -q                      Only output flamegraph URI
+  -q | --quiet            Only output flamegraph URI
                           Default: false
 
-  --cmd | -c              Command Mode - A tool set for processing data in
-                          the profile folder.
-
-                          Run 0x --cmd help for a list of commands
+  -s | --silent           Complete silence, 0x will not output anything
+                          Default: false
 
   --timestamp-profiles    Add timestamp to you profile folder's name
 
+  --gen [flags] <file> |  Generate the flamegraph from a stacks.out file. 
+  -g [flags] <file>       Possible flags include --tiers and --langs and --name
+                          Outputs HTML to STDOUT by default, unless --name is set.
+                          
+  --json-stacks           Output a JSON file of stacks as stacks.<pid>.json
+
+
+  --log-output            The logging output stream. Can be stdout or stderr
+                          Default: stderr
